+++ src/main/java/ua/goit/group6/service/impl/StartupServiceImpl.java
@@ -47,7 +47,7 @@ public class StartupServiceImpl extends AbstractBasicServiceImpl<Startup> implem
 
     @Override
     @Transactional(readOnly = true)
-    public List<Startup> getAllByUserIdDesc(Long id) {
+    public List<Startup> getAllByUserIdDesc(int id) {
         return userDao.getById(id)
                 .getStartups()
                 .stream()
